+++ src/routes/+page.svelte
@@ -28,6 +28,7 @@
 	import User from '@/types/user';
 	import UserTable from '@/types/user_table';
 	import RadioGroup from '@cp/Inputs/RadioGroup.svelte';
+	import Tooltip from '@cp/Tooltip.svelte';
 
 	let dialog: Dialog;
 
@@ -45,6 +46,7 @@
 	let password = '';
 
 	let open = false;
+	let new_open = false;
 
 	let openNotifications = false;
 </script>
@@ -377,6 +379,7 @@
 		<div class="h-full w-full container lg:pt-[76px] sm:pt-[60px] pt-[53px]">
 			<div class="h-full relative px-5 pt-5">
 				<DropdownMenu
+					id="0"
 					bind:open
 					transition="scale"
 					class="pb-3"
@@ -397,7 +400,7 @@
 				>
 					<Button
 						on:click={() => (open = !open)}
-						data={{ 'data-ignore-click-outside': true }}
+						data={{ 'data-ignore-click-outside': 0 }}
 						class="group bg-white dark:bg-gray-800 !shadow-lg"
 						color="none"
 						><span class="mx-1">{$LL.DropDown.Name()}</span>
@@ -481,6 +484,26 @@
 	<div
 		class="flex flex-col h-screen justify-center items-center bg-background-reverse snap-start relative"
 	>
+		<Tooltip xDirection="left" yDirection="center">
+			<div
+				class="text-gray-600 transition-colors duration-200 focus:outline-none dark:text-gray-200 dark:hover:text-blue-400 hover:text-blue-500"
+			>
+				<svg
+					xmlns="http://www.w3.org/2000/svg"
+					fill="none"
+					viewBox="0 0 24 24"
+					stroke-width="2"
+					stroke="currentColor"
+					class="w-6 h-6"
+				>
+					<path
+						stroke-linecap="round"
+						stroke-linejoin="round"
+						d="M9.879 7.519c1.171-1.025 3.071-1.025 4.242 0 1.172 1.025 1.172 2.687 0 3.712-.203.179-.43.326-.67.442-.745.361-1.45.999-1.45 1.827v.75M21 12a9 9 0 11-18 0 9 9 0 0118 0zm-9 5.25h.008v.008H12v-.008z"
+					/>
+				</svg>
+			</div>
+		</Tooltip>
 		<RadioGroup
 			items={[
 				{
@@ -512,7 +535,6 @@
 			label={$LL.RadioButton.Question()}
 			name="survey"
 			color="blue"
-			hideInput
 		/>
 	</div>
 </div>